+++ packages/iov-keycontrol/src/userprofile.ts
@@ -3,27 +3,19 @@ import { LevelUp } from "levelup";
 import { ReadonlyDate } from "readonly-date";
 
 import { FullSignature, Nonce, SignedTransaction, TxCodec, UnsignedTransaction } from "@iov/bcp-types";
-import {
-  Argon2id,
-  Argon2idOptions,
-  Random,
-  Slip10RawIndex,
-  Xchacha20poly1305Ietf,
-  Xchacha20poly1305IetfCiphertext,
-  Xchacha20poly1305IetfKey,
-  Xchacha20poly1305IetfMessage,
-  Xchacha20poly1305IetfNonce,
-} from "@iov/crypto";
-import { Encoding } from "@iov/encoding";
+import { Argon2id, Argon2idOptions, Slip10RawIndex } from "@iov/crypto";
+import { Encoding, Int53 } from "@iov/encoding";
 import { DefaultValueProducer, ValueAndUpdates } from "@iov/stream";
 
-import { Keyring, KeyringSerializationString } from "./keyring";
+import { Keyring } from "./keyring";
+import { EncryptedKeyring, KeyringEncryptor } from "./keyringencryptor";
 import { DatabaseUtils } from "./utils";
 import { LocalIdentity, PublicIdentity, Wallet, WalletId } from "./wallet";
 import { Ed25519Wallet } from "./wallets";
 
-const { toAscii, fromBase64, toBase64, fromUtf8, toUtf8, toRfc3339, fromRfc3339 } = Encoding;
+const { toAscii, fromBase64, toBase64, toRfc3339, fromRfc3339 } = Encoding;
 
+const storageKeyFormatVersion = "format_version";
 const storageKeyCreatedAt = "created_at";
 const storageKeyKeyring = "keyring";
 
@@ -63,30 +55,26 @@ export class UserProfile {
     db: LevelUp<AbstractLevelDOWN<string, string>>,
     password: string,
   ): Promise<UserProfile> {
-    // get from storage (raw strings)
-    const createdAtFromStorage = await db.get(storageKeyCreatedAt, { asBuffer: false });
-    const keyringFromStorage = await db.get(storageKeyKeyring, { asBuffer: false });
+    const formatVersion = Int53.fromString(await db.get(storageKeyFormatVersion, { asBuffer: false }));
 
-    // process
-    const encryptionKey = (await Argon2id.execute(
-      password,
-      userProfileSalt,
-      weakPasswordHashingOptions,
-    )) as Xchacha20poly1305IetfKey;
-    const keyringBundle = fromBase64(keyringFromStorage);
-    const keyringNonce = keyringBundle.slice(0, 24) as Xchacha20poly1305IetfNonce;
-    const keyringCiphertext = keyringBundle.slice(24) as Xchacha20poly1305IetfCiphertext;
-    const decrypted = await Xchacha20poly1305Ietf.decrypt(keyringCiphertext, encryptionKey, keyringNonce);
-    const keyringSerialization = fromUtf8(decrypted) as KeyringSerializationString;
+    switch (formatVersion.toNumber()) {
+      case 1:
+        // get from storage (raw strings)
+        const createdAtFromStorage = await db.get(storageKeyCreatedAt, { asBuffer: false });
+        const keyringFromStorage = await db.get(storageKeyKeyring, { asBuffer: false });
 
-    // create objects
-    const createdAt = fromRfc3339(createdAtFromStorage);
-    const keyring = new Keyring(keyringSerialization);
-    return new UserProfile({ createdAt, keyring });
-  }
+        // process
+        const encryptionKey = await Argon2id.execute(password, userProfileSalt, weakPasswordHashingOptions);
+        const encryptedKeyring = fromBase64(keyringFromStorage) as EncryptedKeyring;
+        const keyringSerialization = await KeyringEncryptor.decrypt(encryptedKeyring, encryptionKey);
 
-  private static async makeNonce(): Promise<Xchacha20poly1305IetfNonce> {
-    return (await Random.getBytes(24)) as Xchacha20poly1305IetfNonce;
+        // create objects
+        const createdAt = fromRfc3339(createdAtFromStorage);
+        const keyring = new Keyring(keyringSerialization);
+        return new UserProfile({ createdAt, keyring });
+      default:
+        throw new Error(`Unsupported format version: ${formatVersion.toNumber()}`);
+    }
   }
 
   public readonly createdAt: ReadonlyDate;
@@ -124,24 +112,16 @@ export class UserProfile {
     await DatabaseUtils.clear(db);
 
     // process
-    const encryptionKey = (await Argon2id.execute(
-      password,
-      userProfileSalt,
-      weakPasswordHashingOptions,
-    )) as Xchacha20poly1305IetfKey;
-    const keyringPlaintext = toUtf8(this.keyring.serialize()) as Xchacha20poly1305IetfMessage;
-    const keyringNonce = await UserProfile.makeNonce();
-    const keyringCiphertext = await Xchacha20poly1305Ietf.encrypt(
-      keyringPlaintext,
-      encryptionKey,
-      keyringNonce,
-    );
+    const encryptionKey = await Argon2id.execute(password, userProfileSalt, weakPasswordHashingOptions);
+    const encryptedKeyring = await KeyringEncryptor.encrypt(this.keyring.serialize(), encryptionKey);
 
     // create storage values (raw strings)
+    const formatVersionForStorage = "1";
     const createdAtForStorage = toRfc3339(this.createdAt);
-    const keyringForStorage = toBase64(new Uint8Array([...keyringNonce, ...keyringCiphertext]));
+    const keyringForStorage = toBase64(encryptedKeyring);
 
     // store
+    await db.put(storageKeyFormatVersion, formatVersionForStorage);
     await db.put(storageKeyCreatedAt, createdAtForStorage);
     await db.put(storageKeyKeyring, keyringForStorage);
   }